Move BlendColourMap to its own file
