+++ src/openrct2-ui/drawing/engines/opengl/TextureCache.cpp
@@ -14,6 +14,7 @@
     #include <algorithm>
     #include <openrct2/Diagnostic.h>
     #include <openrct2/core/EnumUtils.hpp>
+    #include <openrct2/drawing/BlendColourMap.h>
     #include <openrct2/drawing/Drawing.h>
     #include <openrct2/interface/Colour.h>
     #include <openrct2/world/Location.hpp>
