+++ src/openrct2/interface/Colour.cpp
@@ -9,12 +9,8 @@
 
 #include "Colour.h"
 
-#include "../SpriteIds.h"
 #include "../core/EnumMap.hpp"
-#include "../core/EnumUtils.hpp"
 #include "../drawing/Drawing.h"
-
-#include <cmath>
 
 using namespace OpenRCT2::Drawing;
 
@@ -95,71 +91,3 @@ namespace OpenRCT2::Colour
     }
 
 } // namespace OpenRCT2::Colour
-
-#ifndef DISABLE_TTF
-static BlendColourMapType BlendColourMap = {};
-
-static bool BlendColourMapInitialised = false;
-
-static PaletteIndex FindClosestPaletteIndex(uint8_t red, uint8_t green, uint8_t blue)
-{
-    PaletteIndex closest = PaletteIndex::pi255;
-    int32_t closestDistance = INT32_MAX;
-
-    for (auto i = PaletteIndex::transparent; i < PaletteIndex::waterWaves0; i = static_cast<PaletteIndex>(EnumValue(i) + 1))
-    {
-        const auto& paletteEntry = gPalette[EnumValue(i)];
-        const int32_t distance = std::pow(paletteEntry.red - red, 2) + std::pow(paletteEntry.green - green, 2)
-            + std::pow(paletteEntry.blue - blue, 2);
-
-        if (distance < closestDistance)
-        {
-            closest = i;
-            closestDistance = distance;
-        }
-    }
-
-    return closest;
-}
-
-static void InitBlendColourMap()
-{
-    for (size_t i = 0; i < kGamePaletteSize; i++)
-    {
-        for (size_t j = i; j < kGamePaletteSize; j++)
-        {
-            uint8_t red = (gPalette[i].red + gPalette[j].red) / 2;
-            uint8_t green = (gPalette[i].green + gPalette[j].green) / 2;
-            uint8_t blue = (gPalette[i].blue + gPalette[j].blue) / 2;
-
-            auto colour = FindClosestPaletteIndex(red, green, blue);
-            BlendColourMap[i][j] = colour;
-            BlendColourMap[j][i] = colour;
-        }
-    }
-    BlendColourMapInitialised = true;
-}
-
-BlendColourMapType* GetBlendColourMap()
-{
-    if (!BlendColourMapInitialised)
-    {
-        InitBlendColourMap();
-    }
-    return &BlendColourMap;
-}
-
-PaletteIndex BlendColours(const PaletteIndex paletteIndex1, const PaletteIndex paletteIndex2)
-{
-    if (!BlendColourMapInitialised)
-    {
-        InitBlendColourMap();
-    }
-    return BlendColourMap[EnumValue(paletteIndex1)][EnumValue(paletteIndex2)];
-}
-#else
-BlendColourMapType* GetBlendColourMap()
-{
-    return nullptr;
-}
-#endif
